move to web3-utils
