+++ packages/web3-utils/src/index.js
@@ -27,7 +27,7 @@ var ethjsUnit = require('ethjs-unit');
 var utils = require('./utils.js');
 var soliditySha3 = require('./soliditySha3.js');
 var randombytes = require('randombytes');
-
+var BN = require('bn.js');
 
 
 /**
@@ -316,6 +316,62 @@ var toChecksumAddress = function (address) {
     return checksumAddress;
 };
 
+/**
+ * Returns -1 if a<b, 1 if a>b; 0 if a == b.
+ * For more details on this type of function, see
+ * developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort
+ *
+ * @method compareBlockNumbers
+ *
+ * @param {String|Number|BN} a
+ *
+ * @param {String|Number|BN} b
+ *
+ * @returns {Number} -1, 0, or 1
+ */
+var compareBlockNumbers = function(a, b) {
+    if (a == b) {
+        return 0;
+    } else if (("genesis" == a || "earliest" == a || 0 == a) && ("genesis" == b || "earliest" ==  b) || 0 == a) {
+        return 0;
+    } else if ("genesis" == a || "earliest" == a) {
+        // b !== a, thus a < b
+        return -1;
+    } else if ("genesis" == b || "earliest" == b) {
+        // b !== a, thus a > b
+        return 1;
+    } else if (a == "latest") {
+        if (b == "pending") {
+            return -1;
+        } else {
+            // b !== ("pending" OR "latest"), thus a > b
+            return 1;
+        }
+    } else if (b === "latest") {
+        if (a == "pending") {
+            return 1;
+        } else {
+            // b !== ("pending" OR "latest"), thus a > b
+            return -1 
+        }
+    } else if (a == "pending") {
+        // b (== OR <) "latest", thus a > b
+        return 1;
+    } else if (b == "pending") {
+        return -1;
+    } else {
+        let bnA = new BN(a);
+        let bnB = new BN(b);
+        if(bnA.lt(bnB)) {
+            return -1;
+        } else if(bnA.eq(bnB)) {
+            return 0;
+        } else {
+            return 1;
+        }
+    }
+};
+
 module.exports = {
     _fireError: _fireError,
     _jsonInterfaceMethodToString: _jsonInterfaceMethodToString,
@@ -379,5 +435,7 @@ module.exports = {
     isContractAddressInBloom: utils.isContractAddressInBloom,
     isTopic: utils.isTopic,
     isTopicInBloom: utils.isTopicInBloom,
-    isInBloom: utils.isInBloom
+    isInBloom: utils.isInBloom,
+
+    compareBlockNumbers: compareBlockNumbers
 };
